task-2117: the GitHub tests workflow gets past its first step on both systems
- Linux: every .sh is now executable in git, and the workflow runs the two
  setup scripts through bash. tools/sqlite-reference.sh answered exit 126 on
  every Linux run.
- Windows: cargo clippy --all-targets --all-features -D warnings is clean on
  the pinned 1.95.0. apply_row carries the room check's answer in Landing
  instead of taking nine parameters; the rest are clippy's own fixes and
  three test-only cases.

+++ crates/inillucent-compat/tests/delete_order.rs
@@ -174,7 +174,7 @@ fn build(connection: &Connection<'_>, rows: i64, period: i64) {
 #[test]
 fn a_delete_in_index_order_reads_each_leaf_about_once() {
     let path = scratch("index-order");
-    let database = Database::open_at(&path, 4_096, inillucent_engine::DEFAULT_FRAMES as usize)
+    let database = Database::open_at(&path, 4_096, inillucent_engine::DEFAULT_FRAMES)
         .expect("the database opens");
     let connection = database.session();
     exec(&connection, &format!("PRAGMA cache_size = {POOL_PAGES}"));
@@ -253,7 +253,7 @@ fn a_delete_in_index_order_reads_each_leaf_about_once() {
 #[test]
 fn a_trigger_and_returning_see_the_query_order() {
     let path = scratch("observable");
-    let database = Database::open_at(&path, 4_096, inillucent_engine::DEFAULT_FRAMES as usize)
+    let database = Database::open_at(&path, 4_096, inillucent_engine::DEFAULT_FRAMES)
         .expect("the database opens");
     let connection = database.session();
     build(&connection, 60, 7);
